Reintroduce Django-1.6 support.

+++ run_tests.py
@@ -5,6 +5,7 @@ os.environ['DJANGO_SETTINGS_MODULE'] = 'tests.settings'
 import django
 from django.core.management import call_command
 
+names_prefix = 'tests.tests' if django.VERSION >= (1, 6) else 'tests'
 names = next((a for a in sys.argv[1:] if not a.startswith('-')), None)
 if not names:
     names = 'tests'
@@ -13,15 +14,16 @@ elif re.search(r'^\d+', names):
 elif not names.startswith('tests.'):
     names = 'tests.tests.' + names
 
-
-django.setup()
-
+if hasattr(django, 'setup'):
+    django.setup()
 
 # NOTE: we create migrations each time  since they depend on type of database,
 #       python and django versions
 try:
-    shutil.rmtree('tests/migrations', True)
-    call_command('makemigrations', 'tests', verbosity=2 if '-v' in sys.argv else 0)
+    if django.VERSION >= (1, 7):
+        shutil.rmtree('tests/migrations', True)
+        call_command('makemigrations', 'tests', verbosity=2 if '-v' in sys.argv else 0)
     call_command('test', names, failfast='-x' in sys.argv, verbosity=2 if '-v' in sys.argv else 1)
 finally:
-    shutil.rmtree('tests/migrations')
+    if django.VERSION >= (1, 7):
+        shutil.rmtree('tests/migrations')
